requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pyyaml==6.0.2
httpx==0.28.1
qdrant-client==1.12.1
python-multipart==0.0.20
